requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
greenlet==3.3.2
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.7
